requirements.txt
# Core Agent Framework
strands-agents>=1.13.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-aws>=0.1.0
bedrock-agentcore[strands-agents]>=0.1.0

# Document Processing
docling>=2.58.0
pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
pillow==11.0.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Web Framework
streamlit>=1.31.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
plotly>=5,<6

# Data Models
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
structlog>=24.1.0
python-multipart>=0.0.6
jinja2==3.1.6
markdown==3.7

# Date/Time
python-dateutil==2.9.0.post0

# JSON/YAML
pyyaml==6.0.2
# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Optional: AgentCore
bedrock-agentcore-starter-toolkit>=0.1.0

# Additional dependencies for stability
typing-extensions==4.13.2
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
urllib3==2.2.3